lart command moved to own package
