+++ src/main/java/com/fossgalaxy/pircbotx/backends/irc/IrcService.java
@@ -6,7 +6,7 @@ import com.fossgalaxy.pircbotx.backends.BotService;
 import com.fossgalaxy.pircbotx.backends.ChannelService;
 import com.fossgalaxy.pircbotx.backends.UserService;
 import com.fossgalaxy.pircbotx.commandprocessor.CommandProcessor;
-import com.fossgalaxy.pircbotx.listeners.JoinsListener;
+import com.fossgalaxy.pircbotx.commands.joins.JoinsListener;
 import com.fossgalaxy.pircbotx.listeners.LinesListener;
 import com.google.inject.Inject;
 import com.google.inject.Injector;
