+++ src/main/java/com/fossgalaxy/pircbotx/commands/joins/JoinsListener.java
@@ -16,9 +16,8 @@
  * You should have received a copy of the GNU General Public License along with
  * uc_pircbotx. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.fossgalaxy.pircbotx.listeners;
+package com.fossgalaxy.pircbotx.commands.joins;
 
-import com.fossgalaxy.pircbotx.data.db.JoinModel;
 import com.google.inject.Inject;
 import org.pircbotx.User;
 import org.pircbotx.hooks.ListenerAdapter;
